+++ src/pages/benchmarks/styles.module.css
@@ -1,0 +1,4 @@
+.container {
+  width: 90vw;
+  align-self: center;
+}
